refactor: extract composite typography semantic tokens
