+++ OUDS/Core/OUDS/Sources/OUDSTheme/OUDSTheme+SemanticTokens/OUDSTheme+ElevationCompositeSemanticTokens.swift
@@ -15,13 +15,18 @@ import Foundation
 import OUDSTokensRaw
 import OUDSTokensSemantic
 
-/// Defines basic values common to all themes for `CompositeElevationSemanticTokens`.
+// ଘ( ･ω･)_/ﾟ･:*:･｡☆
+// [File to generate with the tokenator with Figma able to output composites and tokenatoer able to manage them]
+// WARNING: Not synchronized anymore with the Figjam / Figma by developers team
+// Create an issue for update https://github.com/Orange-OpenSource/ouds-ios/issues/new?template=token_update.yml
+
+/// Defines basic values common to all themes for `ElevationCompositeSemanticTokens`.
 /// These values can be overriden inside `OUDSTheme` subclasses (in extensions or not, in the same module or not) thanks to the `@objc open` combination.
 /// The aim of this extensions is to make relationships between all semantic tokens for elevations and associated raw tokens.
 /// `OUDSTheme` can be seen as a kind of "abstract class" in _object oriented paradigm_.
 ///
 /// It defines in fact box shadows effects.
-extension OUDSTheme: CompositeElevationSemanticTokens {
+extension OUDSTheme: ElevationCompositeSemanticTokens {
 
     // MARK: Semantic token - Elevation - Box shadow
