+++ OUDS/Core/Tokens/RawTokens/Sources/Values/ElevationRawTokens+Composites.swift
@@ -12,8 +12,8 @@
 //
 
 // ଘ( ･ω･)_/ﾟ･:*:･｡☆
-// [File to generate with the tokenator]
-// WARNING: Not synchronized anymore with the Figjam / Figma by developer team
+// [File to generate with the tokenator with Figma able to output composites and tokenatoer able to manage them]
+// WARNING: Not synchronized anymore with the Figjam / Figma by developers team
 // Create an issue for update https://github.com/Orange-OpenSource/ouds-ios/issues/new?template=token_update.yml
 
 // swiftlint:disable missing_docs
