+++ OUDS/Core/Tokens/SemanticTokens/Sources/Values/ElevationSemanticTokens.swift
@@ -19,10 +19,10 @@
 // swiftlint:disable missing_docs
 
 /// This is a group of semantic tokens for **elevations**.
-/// It defines all elevation semantic tokens a theme must have. (`ElevationXSemanticToken`,`ElevationYSemanticToken`,
+/// It defines all elevation semantic tokens a theme must have (`ElevationXSemanticToken`,`ElevationYSemanticToken`,
 /// `ElevationBlurSemanticToken` and `ElevationColorSemanticToken`).
 ///
-/// However the composite tokens (here for bow shadows) are defined in `CompositeElevationSemanticTokens` because the *tokenator*
+/// However the composite tokens (here for bow shadows) are defined in `ElevationCompositeSemanticTokens` because the *tokenator*
 /// is not able to generate them yet, and they must be defined elsewhere to not be deleted.
 ///
 /// In the future this file should be generated by a parser converting Figma JSON to Swift code.
