+++ OUDS/Core/Tokens/SemanticTokens/Sources/Values/TypographySemanticTokens.swift
@@ -22,6 +22,9 @@
 /// It defines all typography semantic tokens a theme must have
 /// (`TypographyFontFamilySemanticToken`, `TypographyFontWeightSemanticToken`, `TypographyFontSizeSemanticToken`, `TypographyFontLineHeightSemanticToken`).
 ///
+/// However the composite tokens (here the ones gathering each type of semantic token here) are defined in `TypographyCompositeSemanticTokens` because the *tokenator*
+/// is not able to generate them yet, and they must be defined elsewhere to not be deleted.
+///
 /// In the future this file should be generated by a parser converting Figma JSON to Swift code.
 public protocol TypographySemanticTokens {
 
@@ -106,44 +109,6 @@ public protocol TypographySemanticTokens {
     var fontLetterSpacingLabelSmall: MultipleFontLetterSpacingTokens { get }
     var fontLetterSpacingCodeMedium: MultipleFontLetterSpacingTokens { get }
     var fontLetterSpacingCodeSmall: MultipleFontLetterSpacingTokens { get }
-
-    // MARK: - Semantic tokens - Typography - Composites - Display
-
-    var typeDisplayLarge: MultipleTypographyTokens { get }
-    var typeDisplayMedium: MultipleTypographyTokens { get }
-    var typeDisplaySmall: MultipleTypographyTokens { get }
-
-    // MARK: - Semantic tokens - Typography - Composites - Heading
-
-    var typeHeadingXLarge: MultipleTypographyTokens { get }
-    var typeHeadingLarge: MultipleTypographyTokens { get }
-    var typeHeadingMedium: MultipleTypographyTokens { get }
-    var typeHeadingSmall: MultipleTypographyTokens { get }
-
-    // MARK: - Semantic tokens - Typography - Composites - Body
-
-    var typeBodyDefaultLarge: MultipleTypographyTokens { get }
-    var typeBodyDefaultMedium: MultipleTypographyTokens { get }
-    var typeBodyDefaultSmall: MultipleTypographyTokens { get }
-    var typeBodyStrongLarge: MultipleTypographyTokens { get }
-    var typeBodyStrongMedium: MultipleTypographyTokens { get }
-    var typeBodyStrongSmall: MultipleTypographyTokens { get }
-
-    // MARK: - Semantic tokens - Typography - Composites - Label
-
-    var typeLabelDefaultXLarge: MultipleTypographyTokens { get }
-    var typeLabelDefaultLarge: MultipleTypographyTokens { get }
-    var typeLabelDefaultMedium: MultipleTypographyTokens { get }
-    var typeLabelDefaultSmall: MultipleTypographyTokens { get }
-    var typeLabelStrongXLarge: MultipleTypographyTokens { get }
-    var typeLabelStrongLarge: MultipleTypographyTokens { get }
-    var typeLabelStrongMedium: MultipleTypographyTokens { get }
-    var typeLabelStrongSmall: MultipleTypographyTokens { get }
-
-    // MARK: - Semantic tokens - Typography - Composites - Code
-
-    var typeCodeMedium: MultipleTypographyTokens { get }
-    var typeCodeSmall: MultipleTypographyTokens { get }
 }
 
 // swiftlint:enable missing_docs
